% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FernandaSanchezRSeries03-package.R
\docType{package}
\name{FernandaSanchezRSeries03-package}
\alias{FernandaSanchezRSeries03-package}
\alias{FernandaSanchezRSeries03}
\title{FernandaSanchezRSeries03}
\description{
\code{FernandaSanchezRSeries03} is a dummy package containing a simple birth-death
difference equation model.
}
\examples{
\dontrun{
# To run the demo
demo("d0105_run_birth_death", package = "FernandaSanchezRSeries03")
}

}
\author{
Fernanda Sanchez <2794747S@student.gla.ac.uk> (maintainer) \cr
Richard Reeve <richard.reeve@glasgow.ac.uk>
}
