% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestep_stochastic_birth_death.R
\name{timestep_stochastic_birth_death}
\alias{timestep_stochastic_birth_death}
\title{timestep_stochastic_birth_death}
\usage{
timestep_stochastic_birth_death(latest, birth.rate, death.rate, timestep)
}
\arguments{
\item{latest}{a data.frame containing the latest population count and time
(column is 'count' and 'time')}

\item{birth.rate}{the birth rate}

\item{death.rate}{the death rate}

\item{timestep}{the time interval}
}
\value{
Returns a data.frame containing the updated population and time
}
\description{
Run a timestep of a simple stochastic birth-death model
}
