% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestep_deterministic_SIS.R
\name{timestep_deterministic_SIS}
\alias{timestep_deterministic_SIS}
\title{timestep_deterministic_SIS
Run a timestep of a simple deterministic SIS model}
\usage{
timestep_deterministic_SIS(latest, transmission.rate, recovery.rate, timestep)
}
\arguments{
\item{latest}{a data.frame containing the latest infected and susceptible population, as well as time
(column is 'infecteds', 'susceptibles' and 'time')}

\item{transmission.rate}{the transmission rate}

\item{recovery.rate}{the recovery rate}

\item{timestep}{the time interval}
}
\value{
Returns a data.frame containing the updated population of infecteds and susceptibles, as well as time
}
\description{
timestep_deterministic_SIS
Run a timestep of a simple deterministic SIS model
}
