% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestep_stochastic_birth_death.R
\name{timestep_stochastic_birth_death}
\alias{timestep_stochastic_birth_death}
\title{timestep_stochastic_birth_death}
\usage{
timestep_stochastic_birth_death(latest, birth.rate, death.rate)
}
\arguments{
\item{latest}{a data.frame containing the latest population count
(column is 'count')}

\item{birth.rate}{the birth rate}

\item{death.rate}{the death rate}
}
\value{
Returns a data.frame containing the updated population
}
\description{
Run one step of a simple deterministic exponential birth-death model
}
