% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestep_stochastic_SIS.R
\name{timestep_stochastic_SIS}
\alias{timestep_stochastic_SIS}
\title{timestep_stochastic_SIS
Run one step of a simple deterministic SIS model}
\usage{
timestep_stochastic_SIS(latest, transmission.rate, recovery.rate, timestep)
}
\arguments{
\item{latest}{a data.frame containing the latest population count
(column is 'count')}

\item{transmission.rate}{the birth rate}

\item{recovery.rate}{the death rate}

\item{timestep}{}
}
\value{
Returns a data.frame containing the updated population
}
\description{
timestep_stochastic_SIS
Run one step of a simple deterministic SIS model
}
